requirements.txt
aioquic==1.2.0
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==3.0.2
mitmproxy==11.0.0
mitmproxy-windows==0.9.3
mitmproxy_rs==0.9.3
msgpack==1.1.0
passlib==1.7.4
protobuf==5.28.2
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydivert==2.1.0
pylsqpack==0.3.18
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyperclip==1.9.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
service-identity==24.2.0
sortedcontainers==2.4.0
tornado==6.4.1
typing_extensions==4.12.2
urwid==2.6.15
wcwidth==0.2.13
Werkzeug==3.0.6
wsproto==1.2.0
zstandard==0.23.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
python-dotenv==1.0.1
redis==5.0.1
MetaTrader5==5.0.4424
numpy==1.24.3
requests==2.31.0
